refactor: getTimeAgo 및 formatPrice 유틸리티 함수 통합 및 불필요한 파일 삭제
