--- apps/client/src/shared/utils/format-price.ts
@@ -1,3 +1,0 @@
-export const formatPrice = (price: number): string => {
-  return price.toLocaleString('ko-KR');
-};
