+++ apps/client/src/widgets/mypage/components/post-preview/post-preview.tsx
@@ -1,7 +1,7 @@
 import { Content, Title } from '@bds/ui';
 import { Icon } from '@bds/ui/icons';
 
-import { getTimeAgo } from '@shared/utils/get-time-ago';
+import { getTimeAgo } from '@shared/utils/utils';
 
 import * as styles from './post-preview.css';
